demos/01-openclaw-basic/requirements.txt
avp-sdk>=0.1.0
anthropic>=0.18.0
